Rename browserprocess to osext
This rename better defines what the package does.

Resolves: https://github.com/grafana/xk6-browser/pull/755#discussion_r1098860638
